backend/requirements.txt
# Flask==2.3.2
# Werkzeug==2.3.6
Flask==3.0.3
Werkzeug==3.0.4
python-dotenv==0.19.2
supabase==1.0.3
Flask-Session==0.8.0
flask-cors==5.0.0